+++ ecd-20200930/alibabacloud_ecd20200930/models/_file_permission_member.py
@@ -13,10 +13,32 @@ class FilePermissionMember(DaraModel):
         expire_time: int = None,
         role_id: str = None,
     ):
+        # The object that you want to grant permissions. The object can be a user or a group.
+        # 
         # This parameter is required.
         self.cds_identity = cds_identity
+        # Indicates whether to disable the permission from users in the subgroup.
         self.disinherit_sub_group = disinherit_sub_group
+        # The expiration time. This value is a UNIX timestamp representing the number of milliseconds that have elapsed since the epoch time January 1, 1970, 00:00:00 UTC
         self.expire_time = expire_time
+        # The role.
+        # 
+        # Valid values:
+        # 
+        # *   SystemFileEditorWithoutShareLink: The role that can edit but cannot share files.
+        # *   SystemFileUploaderAndDownloaderWithShareLink: The role that can upload, download, and share files.
+        # *   SystemFileDownloader: The role that can download files.
+        # *   SystemFileEditorWithoutDelete: The role that can edit but cannot edit files.
+        # *   SystemFileOwner: The role that can collaborate with others on files.
+        # *   SystemFileDownloaderWithShareLink: The role that can download and share files.
+        # *   SystemFileUploaderAndViewer: The role that can preview and upload files.
+        # *   SystemFileViewer: The role that can preview files.
+        # *   SystemFileEditor: The role that can edit files.
+        # *   SystemFileUploaderWithShareLink: The role that can upload and share files.
+        # *   SystemFileUploader: The role that can upload files.
+        # *   SystemFileUploaderAndDownloader: The role that can upload and download files.
+        # *   SystemFileMetaViewer: The role that can view file list.
+        # 
         # This parameter is required.
         self.role_id = role_id
 
@@ -68,8 +90,17 @@ class FilePermissionMemberCdsIdentity(DaraModel):
         id: str = None,
         type: str = None,
     ):
+        # The user ID or a team ID.
+        # 
         # This parameter is required.
         self.id = id
+        # The object type.
+        # 
+        # Valid values:
+        # 
+        # *   IT_Group: team
+        # *   IT_User: user
+        # 
         # This parameter is required.
         self.type = type